Roll back with_transaction abstraction back to context mgr
