+++ tests/test_connection.py
@@ -1,7 +1,4 @@
 import datetime
-import random
-import threading
-import time
 import unittest
 import uuid
 
@@ -20,31 +17,20 @@ import mongoengine.connection
 from mongoengine import (
     DateTimeField,
     Document,
-    IntField,
     StringField,
     connect,
     disconnect_all,
     register_connection,
-    run_in_transaction,
 )
 from mongoengine.connection import (
     DEFAULT_DATABASE_NAME,
     ConnectionFailure,
     _get_connection_settings,
-    _get_session,
     disconnect,
     get_connection,
     get_db,
 )
-from mongoengine.context_managers import switch_db
-from mongoengine.errors import OperationError
 from mongoengine.pymongo_support import PYMONGO_VERSION
-
-from .utils import (
-    MongoDBTestCase,
-    requires_mongodb_gte_40,
-    requires_mongodb_gte_44,
-)
 
 
 def random_str():
@@ -716,279 +702,6 @@ class ConnectionTest(unittest.TestCase):
             == pymongo.common._UUID_REPRESENTATIONS["pythonLegacy"]
         )
         disconnect(rand)
-
-    def test_transaction_usage_raises_exception_when_using_pymong_lt_39(self):
-        if PYMONGO_VERSION < (3, 9):
-            with pytest.raises(OperationError):
-
-                def cb(session):
-                    pass
-
-                run_in_transaction(cb)
-
-
-class TransactionTest(MongoDBTestCase):
-    @requires_mongodb_gte_40
-    def test_updating_a_document_within_a_transaction(self):
-        if PYMONGO_VERSION < (3, 9):
-            pytest.skip("pymongo>=3.9 is required to use transactions")
-
-        connect("mongoenginetest")
-
-        class A(Document):
-            name = StringField()
-
-        A.drop_collection()
-
-        a_doc = A.objects.create(name="a")
-
-        def cb(session):
-            a_doc.update(name="b")
-
-        run_in_transaction(cb)
-        assert "b" == A.objects.get(id=a_doc.id).name
-
-    @requires_mongodb_gte_40
-    def test_update_a_document_within_a_transaction_with_custom_args(self):
-        if PYMONGO_VERSION < (3, 9):
-            pytest.skip("pymongo>=3.9 is required to use transactions")
-
-        connect("mongoenginetest")
-
-        class A(Document):
-            name = StringField()
-
-        A.drop_collection()
-
-        a_doc = A.objects.create(name="a")
-
-        def cb(session, custom_arg, custom_kwarg=""):
-            a_doc.update(name=f"{custom_arg}-{custom_kwarg}")
-
-        run_in_transaction(lambda s: cb(s, "a", custom_kwarg="b"))
-        assert "a-b" == A.objects.get(id=a_doc.id).name
-
-    @requires_mongodb_gte_40
-    def test_transaction_updates_across_databases(self):
-        if PYMONGO_VERSION < (3, 9):
-            pytest.skip("pymongo>=3.9 is required to use transactions")
-
-        connect("mongoenginetest")
-        connect("test2", "test2")
-
-        class A(Document):
-            name = StringField()
-
-        A.objects.all().delete()
-        a_doc = A.objects.create(name="a")
-
-        class B(Document):
-            meta = {"db_alias": "test2"}
-            name = StringField()
-
-        B.objects.all().delete()
-        b_doc = B.objects.create(name="b")
-
-        def cb(session):
-            a_doc.update(name="a2")
-            b_doc.update(name="b2")
-
-        run_in_transaction(cb)
-
-        assert "a2" == A.objects.get(id=a_doc.id).name
-        assert "b2" == B.objects.get(id=b_doc.id).name
-
-    @requires_mongodb_gte_44
-    def test_collection_creation_via_upserts_across_databases_in_transaction(self):
-        if PYMONGO_VERSION < (3, 9):
-            pytest.skip("pymongo>=3.9 is required to use transactions")
-
-        connect("mongoenginetest")
-        connect("test2", "test2")
-
-        class A(Document):
-            name = StringField()
-
-        A.drop_collection()
-
-        a_doc = A.objects.create(name="a")
-
-        class B(Document):
-            meta = {"db_alias": "test2"}
-            name = StringField()
-
-        B.drop_collection()
-
-        b_doc = B.objects.create(name="b")
-
-        def cb(session):
-            a_doc.update(name="a3")
-            with switch_db(A, "test2"):
-                a_doc.update(name="a4", upsert=True)
-                b_doc.update(name="b3")
-            b_doc.update(name="b4")
-
-        run_in_transaction(cb)
-
-        assert "a3" == A.objects.get(id=a_doc.id).name
-        assert "b4" == B.objects.get(id=b_doc.id).name
-        with switch_db(A, "test2"):
-            assert "a4" == A.objects.get(id=a_doc.id).name
-
-    @requires_mongodb_gte_40
-    def test_an_exception_raised_in_transactions_across_databases_rolls_back_updates(
-        self,
-    ):
-        if PYMONGO_VERSION < (3, 9):
-            pytest.skip("pymongo>=3.9 is required to use transactions")
-
-        connect("mongoenginetest")
-        connect("test2", "test2")
-
-        class A(Document):
-            name = StringField()
-
-        A.drop_collection()
-        with switch_db(A, "test2"):
-            A.drop_collection()
-
-        a_doc = A.objects.create(name="a")
-
-        class B(Document):
-            meta = {"db_alias": "test2"}
-            name = StringField()
-
-        B.drop_collection()
-
-        b_doc = B.objects.create(name="b")
-
-        def cb(session):
-            a_doc.update(name="a3")
-            with switch_db(A, "test2"):
-                a_doc.update(name="a4", upsert=True)
-                b_doc.update(name="b3")
-                b_doc.update(name="b4")
-            raise Exception
-
-        try:
-            run_in_transaction(cb)
-        except Exception:
-            pass
-
-        assert "a" == A.objects.get(id=a_doc.id).name
-        assert "b" == B.objects.get(id=b_doc.id).name
-        with switch_db(A, "test2"):
-            assert 0 == A.objects.all().count()
-
-    @requires_mongodb_gte_40
-    def test_nested_transactions_create_and_release_sessions_accordingly(self):
-        # NOTE: callbacks initiating new transactions goes against pymongo recommended usage!
-        # This is simply a sanity check and should not be considered suggested usage.
-        if PYMONGO_VERSION < (3, 9):
-            pytest.skip("pymongo>=3.9 is required to use transactions")
-
-        connect("mongoenginetest")
-
-        def cb1(s1):
-            def cb2(s2):
-                assert s1 != s2
-                assert s2 == _get_session()
-
-            run_in_transaction(cb2)
-            assert s1 == _get_session()
-
-        run_in_transaction(cb1)
-        assert _get_session() is None
-
-    @requires_mongodb_gte_40
-    def test_thread_safety_of_transactions(self):
-        """
-        Make sure transactions don't step over each other. Each
-        session should be isolated to each thread. If this is the
-        case, then no amount of runtime variability should have
-        an effect on the output.
-
-        This test sets up 10 records, each with an integer field
-        of value 0 - 9.
-
-        We then spin up 10 threads and attempt to update a target
-        record by multiplying it's integer value by 10. Then, if
-        the target record is even, throw an exception, which
-        should then roll the transaction back. The odd rows always
-        succeed.
-
-        If the sessions are properly thread safe, we should ALWAYS
-        net out with the following sum across the integer fields
-        of the 10 records:
-
-        0 + 10 + 2 + 30 + 4 + 50 + 6 + 70 + 8 + 90 = 270
-        """
-        if PYMONGO_VERSION < (3, 9):
-            pytest.skip("pymongo>=3.9 is required to use transactions")
-
-        connect("mongoenginetest")
-
-        class A(Document):
-            i = IntField()
-
-        # Ensure the collection is created
-        A.objects.create(i=0)
-
-        class TestExc(Exception):
-            pass
-
-        def thread_fn(idx):
-            # Open the transaction at some unknown interval
-            time.sleep(random.uniform(0.01, 0.1))
-            try:
-
-                def cb(session):
-                    a = A.objects.get(i=idx)
-                    a.i = idx * 10
-                    # Save at some unknown interval
-                    time.sleep(random.uniform(0.01, 0.1))
-                    a.save()
-
-                    # Force roll backs for the even runs...
-                    if idx % 2 == 0:
-                        raise TestExc
-
-                run_in_transaction(cb)
-            except TestExc:
-                pass
-
-        for r in range(10):
-            """
-            Threads & randomization are tricky - run it multiple times
-            """
-
-            # Clear out the collection for a fresh run
-            A.objects.all().delete()
-
-            # Prepopulate the data for reads
-            thread_count = 10
-            for i in range(thread_count):
-                A.objects.create(i=i)
-
-            # Prime the threads
-            threads = [
-                threading.Thread(target=thread_fn, args=(i,))
-                for i in range(thread_count)
-            ]
-            for t in threads:
-                t.start()
-            for t in threads:
-                t.join()
-
-            # Check the sum
-            expected_sum = 0
-            for i in range(thread_count):
-                if i % 2 == 0:
-                    expected_sum += i
-                else:
-                    expected_sum += i * 10
-            assert expected_sum == 270
-            assert expected_sum == sum(a.i for a in A.objects.all())
 
 
 if __name__ == "__main__":
