+++ tests/test_context_managers.py
@@ -1,4 +1,5 @@
 import random
+import threading
 import time
 import unittest
 from threading import Thread
@@ -7,18 +8,24 @@ import pytest
 from bson import DBRef
 
 from mongoengine import *
-from mongoengine.connection import get_db
+from mongoengine.connection import _get_session, get_db
 from mongoengine.context_managers import (
     no_dereference,
     no_sub_classes,
     query_counter,
+    run_in_transaction,
     set_read_write_concern,
     set_write_concern,
     switch_collection,
     switch_db,
 )
 from mongoengine.pymongo_support import count_documents
-from tests.utils import MongoDBTestCase
+
+from .utils import (
+    MongoDBTestCase,
+    requires_mongodb_gte_40,
+    requires_mongodb_gte_44,
+)
 
 
 class TestableThread(Thread):
@@ -481,6 +488,302 @@ class TestContextManagers(MongoDBTestCase):
                 db.system.indexes.find_one()
             )  # queries on db.system.indexes are ignored as well
             assert q == 1
+
+    @requires_mongodb_gte_40
+    def test_updating_a_document_within_a_transaction(self):
+        connect("mongoenginetest")
+
+        class A(Document):
+            name = StringField()
+
+        A.drop_collection()
+
+        a_doc = A.objects.create(name="a")
+
+        with run_in_transaction():
+            a_doc.update(name="b")
+            assert "b" == A.objects.get(id=a_doc.id).name
+
+    @requires_mongodb_gte_40
+    def test_transaction_updates_across_databases(self):
+        connect("mongoenginetest")
+        connect("test2", "test2")
+
+        class A(Document):
+            name = StringField()
+
+        A.objects.all().delete()
+        a_doc = A.objects.create(name="a")
+
+        class B(Document):
+            meta = {"db_alias": "test2"}
+            name = StringField()
+
+        B.objects.all().delete()
+        b_doc = B.objects.create(name="b")
+
+        with run_in_transaction():
+            a_doc.update(name="a2")
+            b_doc.update(name="b2")
+
+        assert "a2" == A.objects.get(id=a_doc.id).name
+        assert "b2" == B.objects.get(id=b_doc.id).name
+
+    @requires_mongodb_gte_44
+    def test_collection_creation_via_upserts_across_databases_in_transaction(self):
+        connect("mongoenginetest")
+        connect("test2", "test2")
+
+        class A(Document):
+            name = StringField()
+
+        A.drop_collection()
+
+        a_doc = A.objects.create(name="a")
+
+        class B(Document):
+            meta = {"db_alias": "test2"}
+            name = StringField()
+
+        B.drop_collection()
+
+        b_doc = B.objects.create(name="b")
+
+        with run_in_transaction():
+            a_doc.update(name="a3")
+            with switch_db(A, "test2"):
+                a_doc.update(name="a4", upsert=True)
+                b_doc.update(name="b3")
+            b_doc.update(name="b4")
+
+        assert "a3" == A.objects.get(id=a_doc.id).name
+        assert "b4" == B.objects.get(id=b_doc.id).name
+        with switch_db(A, "test2"):
+            assert "a4" == A.objects.get(id=a_doc.id).name
+
+    @requires_mongodb_gte_40
+    def test_an_exception_raised_in_transactions_across_databases_rolls_back_updates(
+        self,
+    ):
+        connect("mongoenginetest")
+        connect("test2", "test2")
+
+        class A(Document):
+            name = StringField()
+
+        A.drop_collection()
+        with switch_db(A, "test2"):
+            A.drop_collection()
+
+        a_doc = A.objects.create(name="a")
+
+        class B(Document):
+            meta = {"db_alias": "test2"}
+            name = StringField()
+
+        B.drop_collection()
+
+        b_doc = B.objects.create(name="b")
+
+        try:
+            with run_in_transaction():
+                a_doc.update(name="a3")
+                with switch_db(A, "test2"):
+                    a_doc.update(name="a4", upsert=True)
+                    b_doc.update(name="b3")
+                    b_doc.update(name="b4")
+                raise Exception
+        except Exception:
+            pass
+
+        assert "a" == A.objects.get(id=a_doc.id).name
+        assert "b" == B.objects.get(id=b_doc.id).name
+        with switch_db(A, "test2"):
+            assert 0 == A.objects.all().count()
+
+    @requires_mongodb_gte_40
+    def test_exception_in_child_of_a_nested_transaction_rolls_parent_back(self):
+        connect("mongoenginetest")
+
+        class A(Document):
+            name = StringField()
+
+        A.drop_collection()
+        a_doc = A.objects.create(name="a")
+
+        class B(Document):
+            name = StringField()
+
+        B.drop_collection()
+        b_doc = B.objects.create(name="b")
+
+        try:
+            with run_in_transaction():
+                a_doc.update(name="trx-parent")
+                with run_in_transaction():
+                    b_doc.update(name="trx-child")
+                    raise Exception
+        except Exception:
+            pass
+
+        assert "a" == A.objects.get(id=a_doc.id).name
+        assert "b" == B.objects.get(id=b_doc.id).name
+
+    @requires_mongodb_gte_40
+    def test_exception_in_parent_of_nested_transaction_after_child_completed_only_rolls_parent_back(
+        self,
+    ):
+        connect("mongoenginetest")
+
+        class A(Document):
+            name = StringField()
+
+        A.drop_collection()
+        a_doc = A.objects.create(name="a")
+
+        class B(Document):
+            name = StringField()
+
+        B.drop_collection()
+        b_doc = B.objects.create(name="b")
+
+        class TestExc(Exception):
+            pass
+
+        def run_tx():
+            try:
+                with run_in_transaction():
+                    a_doc.update(name="trx-parent")
+                    with run_in_transaction():
+                        b_doc.update(name="trx-child")
+                    raise TestExc
+            except TestExc:
+                pass
+            except OperationError as op_failure:
+                """
+                See thread safety test below for more details about TransientTransctionError handling
+                """
+                if "TransientTransactionError" in str(op_failure):
+                    run_tx()
+                else:
+                    raise op_failure
+
+        run_tx()
+        assert "a" == A.objects.get(id=a_doc.id).name
+        assert "trx-child" == B.objects.get(id=b_doc.id).name
+
+    @requires_mongodb_gte_40
+    def test_nested_transactions_create_and_release_sessions_accordingly(self):
+        connect("mongoenginetest")
+        with run_in_transaction():
+            s1 = _get_session()
+            with run_in_transaction():
+                s2 = _get_session()
+                assert s1 != s2
+                with run_in_transaction():
+                    pass
+                assert s2 == _get_session()
+            assert s1 == _get_session()
+        assert _get_session() is None
+
+    @requires_mongodb_gte_40
+    def test_thread_safety_of_transactions(self):
+        """
+        Make sure transactions don't step over each other. Each
+        session should be isolated to each thread. If this is the
+        case, then no amount of runtime variability should have
+        an effect on the output.
+
+        This test sets up 10 records, each with an integer field
+        of value 0 - 9.
+
+        We then spin up 10 threads and attempt to update a target
+        record by multiplying it's integer value by 10. Then, if
+        the target record is even, throw an exception, which
+        should then roll the transaction back. The odd rows always
+        succeed.
+
+        If the sessions are properly thread safe, we should ALWAYS
+        net out with the following sum across the integer fields
+        of the 10 records:
+
+        0 + 10 + 2 + 30 + 4 + 50 + 6 + 70 + 8 + 90 = 270
+        """
+        connect("mongoenginetest")
+
+        class A(Document):
+            i = IntField()
+
+        # Ensure the collection is created
+        A.objects.create(i=0)
+
+        class TestExc(Exception):
+            pass
+
+        def thread_fn(idx):
+            # Open the transaction at some unknown interval
+            time.sleep(random.uniform(0.01, 0.1))
+            try:
+                with run_in_transaction():
+                    a = A.objects.get(i=idx)
+                    a.i = idx * 10
+                    # Save at some unknown interval
+                    time.sleep(random.uniform(0.01, 0.1))
+                    a.save()
+
+                    # Force roll backs for the even runs...
+                    if idx % 2 == 0:
+                        raise TestExc
+            except TestExc:
+                pass
+            except pymongo.errors.OperationFailure as op_failure:
+                """
+                If there's a TransientTransactionError, retry - the lock could not be acquired.
+
+                Per MongoDB docs: The core transaction API does not incorporate retry logic for
+                "TransientTransactionError". To handle "TransientTransactionError", applications
+                should explicitly incorporate retry logic for the error.
+
+                See: https://www.mongodb.com/docs/manual/core/transactions-in-applications/#-transienttransactionerror-
+                """
+                error_labels = op_failure.details.get("errorLabels", [])
+                if "TransientTransactionError" in error_labels:
+                    thread_fn(idx)
+                else:
+                    raise op_failure
+
+        for r in range(10):
+            """
+            Threads & randomization are tricky - run it multiple times
+            """
+
+            # Clear out the collection for a fresh run
+            A.objects.all().delete()
+
+            # Prepopulate the data for reads
+            thread_count = 10
+            for i in range(thread_count):
+                A.objects.create(i=i)
+
+            # Prime the threads
+            threads = [
+                threading.Thread(target=thread_fn, args=(i,))
+                for i in range(thread_count)
+            ]
+            for t in threads:
+                t.start()
+            for t in threads:
+                t.join()
+
+            # Check the sum
+            expected_sum = 0
+            for i in range(thread_count):
+                if i % 2 == 0:
+                    expected_sum += i
+                else:
+                    expected_sum += i * 10
+            assert expected_sum == 270
+            assert expected_sum == sum(a.i for a in A.objects.all())
 
 
 if __name__ == "__main__":
